requirements.txt
Flask
pandas
numpy
matplotlib
gunicorn
scikit-learn
python-dotenv
geopy